requirements.txt
ansible>=7.4
ansible-core>=2.14
ansible-lint>=6.14.0
molecule>=4.0.0,<6.0.0
molecule-plugins[docker]>=23.0.0
pylint>=2.17.0
pytest-testinfra>=7.0.0
yamllint>=1.30.0